change GrainStateHelper class to a separated file
